+++ packages/components/src/AudioPlayer/index.tsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useState, useCallback } from 'react';
-import { ButtonBar, OrderedButton, Slider } from '@jpmorganchase/uitk-lab';
+
+import { ButtonBar, OrderedButton, Slider } from '@salt-ds/lab';
 import { Icon } from '../Icon';
 
 import { Caption3, Caption6 } from '../Typography';