+++ packages/site-components/src/Breadcrumbs/Breadcrumb.tsx
@@ -8,7 +8,7 @@ export interface BreadcrumbProps {
   overflowLabel?: string;
 }
 
-// TODO replace with Odyssey Breadcrumb when it supports an API that can customize Links
+// TODO replace with Salt Breadcrumb when it supports an API that can customize Links
 export const Breadcrumb = forwardRef<HTMLLinkElement, BreadcrumbProps>(function Breadcrumb(
   { children, isCurrentLevel, ...props },
   ref