+++ packages/site-middleware/src/compileMdx.ts
@@ -1,6 +1,6 @@
 import { serialize } from 'next-mdx-remote/serialize';
 import rehypeSlug from 'rehype-slug';
-import { codeBlocks } from '../plugins/codeBlocks';
+import { codeBlocks } from './plugins/codeBlocks.js';
 
 export async function compileMDX(
   content,