+++ packages/standard-generator/src/templates/next.config.js
@@ -34,5 +34,19 @@ module.exports = {
     }
     return config;
   },
-  env: {}
+  env: {},
+  async redirects() {
+    return [
+      {
+        source: '/',
+        destination: '/mosaic/index',
+        permanent: true
+      },
+      {
+        source: '/mosaic',
+        destination: '/mosaic/index',
+        permanent: true
+      }
+    ];
+  }
 };